requirements.txt
openai==1.52.0
httpx==0.27.2
python-dotenv==1.0.1
pyyaml==6.0.2
Jinja2==3.1.4
GitPython==3.1.43
markdown==3.7
unidecode==1.3.8